refactor delete-with-confirmation-icon into a separate element and use it in job-row and job-defenition row
